requirements.txt
flask
flask-socketio
watchdog
numpy 
PyWavelets 
pycryptodome
opencv-python-headless
python-dotenv